+++ outputmarkdown.h
@@ -39,6 +39,8 @@ void toMarkdown(const XmlElement *root_elem,
                 bool create_nav_panel,
                 bool tag_for_each_prefix,
                 bool tag_for_each_suffix,
-                bool graph_connections);
+                bool graph_connections,
+                bool mark_dice_rolls,
+                bool mark_html_dice_rolls);
 
 #endif // OUTPUTMARKDOWN_H